Closes #20: Refatorando a TodoPage para separar as funcionalidades em widgets
